+++ src/lib/config.ts
@@ -46,9 +46,9 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 
 export const SDK_METADATA = {
   language: "typescript",
-  openapiDocVersion: "4.17.0-alpha.1772253819150-d3f36156",
-  sdkVersion: "0.6.0-alpha.31",
-  genVersion: "2.846.1",
+  openapiDocVersion: "4.17.0-alpha.1772665653996-3f22900e",
+  sdkVersion: "0.6.0-alpha.32",
+  genVersion: "2.849.3",
   userAgent:
-    "speakeasy-sdk/typescript 0.6.0-alpha.31 2.846.1 4.17.0-alpha.1772253819150-d3f36156 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.6.0-alpha.32 2.849.3 4.17.0-alpha.1772665653996-3f22900e cribl-control-plane",
 } as const;